split spec files, fix vdom-diff
